+++ foundation_stdlib/init.lua
@@ -1,7 +1,7 @@
 --
 -- Foundation Standard Library (STDLIB)
 --
-local mod = foundation.new_module("foundation_stdlib", "1.36.0")
+local mod = foundation.new_module("foundation_stdlib", "1.37.0")
 
 mod:require("lib/encoding_tables.lua")
 
@@ -30,6 +30,7 @@ mod:require("lib/color.lua")
 mod:require("lib/waves.lua")
 mod:require("lib/easers.lua")
 mod:require("lib/tweener.lua")
+mod:require("lib/plot.lua")
 mod:require("lib/ansi.lua")
 
 if foundation.com.Luna then