+++ foundation_stdlib/lib/number.lua
@@ -1,11 +1,11 @@
--- @namespace foundation.com
+--- @namespace foundation.com
 local unpack = table.unpack or unpack
 local HEX_TABLE = assert(foundation.com.HEX_TABLE)
 local CROCKFORD_BASE32_ENCODE_TABLE = assert(foundation.com.CROCKFORD_BASE32_ENCODE_TABLE)
 
--- Encode an integer as a big-endian binary, len is the length of the string in bytes
---
--- @spec integer_be_encode(Integer, len: Integer): String
+--- Encode an integer as a big-endian binary, len is the length of the string in bytes
+---
+--- @spec integer_be_encode(Integer, len: Integer): String
 function foundation.com.integer_be_encode(integer, len)
   local result = {}
   for i = 0,len-1 do
@@ -15,9 +15,9 @@ function foundation.com.integer_be_encode(integer, len)
   return string.char(unpack(result))
 end
 
--- Encode an integer as a little-endian binary, len is the length of the string in bytes
---
--- @spec integer_le_encode(Integer, len: Integer): String
+--- Encode an integer as a little-endian binary, len is the length of the string in bytes
+---
+--- @spec integer_le_encode(Integer, len: Integer): String
 function foundation.com.integer_le_encode(integer, len)
   local result = {}
   for i = 1,len do
@@ -27,12 +27,12 @@ function foundation.com.integer_le_encode(integer, len)
   return string.char(unpack(result))
 end
 
--- Encode an integer as a big-endian hex string, len is the byte length of the integer
---
--- Usage:
---     integer_hex_be_encode(0xDEADBEEF) -- => DEADBEEF
---
--- @spec integer_hex_be_encode(Integer, len: Integer): String
+--- Encode an integer as a big-endian hex string, len is the byte length of the integer
+---
+--- Usage:
+---     integer_hex_be_encode(0xDEADBEEF) -- => DEADBEEF
+---
+--- @spec integer_hex_be_encode(Integer, len: Integer): String
 function foundation.com.integer_hex_be_encode(integer, len)
   local result = {}
   local byte
@@ -51,9 +51,9 @@ function foundation.com.integer_hex_be_encode(integer, len)
   return table.concat(result)
 end
 
--- Encodes an integer in base16 with big-endian ordering
---
--- @spec integer_base16_be_encode(Integer, len: Integer): String
+--- Encodes an integer in base16 with big-endian ordering
+---
+--- @spec integer_base16_be_encode(Integer, len: Integer): String
 function foundation.com.integer_base16_be_encode(integer, len)
   local segments = len * 2
   local result = {}
@@ -66,9 +66,9 @@ function foundation.com.integer_base16_be_encode(integer, len)
   return table.concat(result)
 end
 
--- Encodes an integer in base16 with little-endian ordering
---
--- @spec integer_base16_le_encode(Integer, len: Integer): String
+--- Encodes an integer in base16 with little-endian ordering
+---
+--- @spec integer_base16_le_encode(Integer, len: Integer): String
 function foundation.com.integer_base16_le_encode(integer, len)
   local segments = len * 2
   local result = {}
@@ -81,7 +81,7 @@ function foundation.com.integer_base16_le_encode(integer, len)
   return table.concat(result)
 end
 
--- @spec integer_crawford_base32_be_encode(Integer, len: Integer, return_table?: Boolean): String
+--- @spec integer_crawford_base32_be_encode(Integer, len: Integer, return_table?: Boolean): String
 function foundation.com.integer_crawford_base32_be_encode(integer, len, return_table)
   local bits = len * 8
   local segments = math.ceil(bits / 5)
@@ -101,7 +101,7 @@ function foundation.com.integer_crawford_base32_be_encode(integer, len, return_t
   end
 end
 
--- @spec integer_crawford_base32_le_encode(Integer, len: Integer, return_table?: Boolean): String
+--- @spec integer_crawford_base32_le_encode(Integer, len: Integer, return_table?: Boolean): String
 function foundation.com.integer_crawford_base32_le_encode(integer, len, return_table)
   local bits = len * 8
   local segments = math.ceil(bits / 5)
@@ -121,7 +121,7 @@ function foundation.com.integer_crawford_base32_le_encode(integer, len, return_t
   end
 end
 
--- @spec number_round(Number): Integer
+--- @spec number_round(Number): Integer
 function foundation.com.number_round(num)
   local floor = math.floor(num)
   local norm = num - floor
@@ -132,15 +132,27 @@ function foundation.com.number_round(num)
   end
 end
 
--- @spec number_lerp(a: Number, b: Number, t: Number): Number
+--- @since "1.37.0"
+--- @spec number_sign(Number): Integer
+function foundation.com.number_sign(num)
+  if num > 0 then
+    return 1
+  elseif num < 0 then
+    return -1
+  else
+    return 0
+  end
+end
+
+--- @spec number_lerp(a: Number, b: Number, t: Number): Number
 function foundation.com.number_lerp(a, b, t)
   return a + (b - a) * t
 end
 
--- Interpolate between 2 numbers by a fixed amount.
--- see `number_lerp` for linear interpolation.
---
--- @spec number_interpolate(a: Number, b: Number, amt: Number): Number
+--- Interpolate between 2 numbers by a fixed amount.
+--- see `number_lerp` for linear interpolation.
+---
+--- @spec number_interpolate(a: Number, b: Number, amt: Number): Number
 function foundation.com.number_interpolate(a, b, amt)
   if a < b then
     return math.min(a + amt, b)
@@ -150,7 +162,7 @@ function foundation.com.number_interpolate(a, b, amt)
   return a
 end
 
--- Deprecated, please use number_interpolate instead.
---
--- @alias number_moveto = number_interpolate
+--- Deprecated, please use number_interpolate instead.
+---
+--- @alias number_moveto = number_interpolate
 foundation.com.number_moveto = foundation.com.number_interpolate